moved files around; ausaxs is actually being invoked now; added automatic download in dev mode
